+++ projects/ngx-parent-injector/child/src/lib/ngx-parent-injector-child.module.ts
@@ -1,30 +1,69 @@
-import { ModuleWithProviders, NgModule } from '@angular/core';
-import { getParentProviders } from './ngx-parent-injector-child-helpers';
-import { NgxParentInjectorChildDirective } from './ngx-parent-injector-child.directive';
+import {
+  ApplicationRef,
+  Inject,
+  ModuleWithProviders,
+  NgModule,
+  NgZone,
+} from '@angular/core';
+import { Subscription } from 'rxjs';
+import {
+  getParentProviders,
+  getParentToken,
+} from './ngx-parent-injector-child-helpers';
 
 export interface NgxParentInjectorChildModuleOptions {
-  testMode?: boolean
+  testMode?: boolean;
 }
 
 /**
  * @private
  * Internal module not meant to be consumed directly. Include `NgxParentInjectorChildModule.forRoot(...)` in your application.
  */
-@NgModule({
-  declarations: [NgxParentInjectorChildDirective],
-  exports: [NgxParentInjectorChildDirective]
-})
-export class NgxParentInjectorChildModuleInternal {}
+@NgModule()
+export class NgxParentInjectorChildModuleInternal {
+  private subscription = new Subscription();
+  public constructor(
+    ngZone: NgZone,
+    @Inject(getParentToken(NgZone)) parentNgZone: NgZone,
+    appRef: ApplicationRef
+  ) {
+    window.addEventListener('beforeunload', () => this.onBeforeUnload());
+
+    // Whenever the parent window's ngZone would trigger change detection,
+    // manually trigger change detection in the child window.
+    this.subscription.add(
+      parentNgZone.onMicrotaskEmpty.subscribe(() => {
+        appRef.tick();
+      })
+    );
+    // Whenever this child window's ngZone would trigger change detection,
+    // trigger another run of the parent window's ngZone. This will trigger
+    // change detection in the parent window, but it will also trigger change
+    // detection in all other child windows too. Using the parent window's
+    // changeDetectorRef here would not propagate changes to other child windows.
+    this.subscription.add(
+      ngZone.onMicrotaskEmpty.subscribe(() => {
+        parentNgZone.run(() => {});
+      })
+    );
+  }
+
+  private onBeforeUnload(): void {
+    this.subscription.unsubscribe();
+  }
+}
 
 export class NgxParentInjectorChildModule {
   /**
    *
    * @param options.testMode default: false. Set to true when running tests.
    */
-  public static forRoot(options?: NgxParentInjectorChildModuleOptions): ModuleWithProviders<NgxParentInjectorChildModuleInternal> {
+  public static forRoot(
+    options?: NgxParentInjectorChildModuleOptions
+  ): ModuleWithProviders<NgxParentInjectorChildModuleInternal> {
     return {
       ngModule: NgxParentInjectorChildModuleInternal,
-      providers: options?.testMode ? [] : getParentProviders()
-    }
+      providers: options?.testMode ? [] : getParentProviders(),
+    };
   }
 }