+++ projects/ngx-parent-injector/child/src/public-api.ts
@@ -1,3 +1,2 @@
 export { getParentToken, suppressWarnings as disableWarnings } from './lib/ngx-parent-injector-child-helpers';
 export * from './lib/ngx-parent-injector-child.module';
-export * from './lib/ngx-parent-injector-child.directive';